Reorganise: move docker commands to `cli._docker_commands`
